tests: move tests to a new file
